Keep item data in Item.getData
Abandon the pattern introduced earlier; a function like getVisibleItems
should not be reactive to both the selection of items and the items'
contents, because changing an item's content should only trigger an
update for that single item. Getting the content should be done in a
separate autorun (by the template/Blaze).